+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 - Write cache initialization happens much faster now, some redundant checks were removed (#3417)
 
 ### Removed
+- Short header support in HEAD's request and response (#3424)
 
 ### Updated
 - `github.com/nspcc-dev/neofs-sdk-go` dependency to `v1.0.0-rc.13.0.20250623124459-a9cfab652dc0` (#3406)